pyproject.toml
[tool.poetry]
name = "scquill"
version = "0.2.6"
description = "Approximate any single cell data set, saving >99% of memory and runtime."
authors = ["Fabio Zanini <fabio.zanini@unsw.edu.au>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
h5py = "^3.10.0"
hdf5plugin = "^4.2.0"
scanpy = "^1.9.6"
pyyaml = "^6.0.1"


[[tool.poetry.source]]
name = "pypi-test"
url = "https://test.pypi.org/legacy/"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-datadir = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
